Move subscription call to graphSubscription
