+++ src/network/graphRequest.js
@@ -135,20 +135,3 @@ export function queryAllOracles(filters, orderBy) {
 export function querySyncInfo() {
   return new GraphRequest('syncInfo').execute();
 }
-
-export function subscribeSyncInfo() {
-  const query = gql`
-    subscription onNewSyncInfo {
-      syncInfo {
-        ${getQueryFields('syncInfo')}
-      }
-    }
-  `;
-
-  client.subscribe({
-    query,
-    variables: {},
-  }).subscribe((response) => {
-    console.log(response);
-  });
-}
